Simplify handling of DisjointClassesAxiom
The axioms are now indexed by producing a list of members (not a set)
and DisjointSubsumer axiom points to the position of the subsumer in
this set (as opposed to the subsumer itself); if two different positions
are produced (even corresponding to the same subsumer), a contradiction
is triggered.

Also, all DisjointSubsumers are now saved in a context to make the
result of the rule application deterministic (and produce all possible
inferences). Previously only at most two different DisjointSubsumers
were saved.