Fix a problem in ConcurrentComputation due to concurrency
The termination flag was not properly synchronized with the interruption
flag, which may cause a situation that submit(I) returns false but the
computation was (not yet) interrupted. Hence
ConcurrentComputationWithInputTest may fail. 

Now do not synchronize these flag but check them both in places which
test if the computation should stop.