New ElkObjectWrapFactory for creating Elk objects from OWLAPI objects
The factory can be used to convert ELK objects created by the reasoner
back to OWLAPI objects since such objects will be created by the factory
"in parallel". The methods of the factory fail if the input Elk objects
are not instances of ElkObjectWrap, i.e., do not have the corresponding
OWLAPI objects attached.