Switch to new method of converting internal inferences to ELK inferences
The indexed objects, conclusion and inferences are now not translated
directly but matched to ELK expressions occurring in the ontology. This
is necessary since indexing of ELK expressions is a lossy operation:
several different ELK expressions may correspond to the same indexed
object. An inference or conclusion "matcher" walks recursively over the
internal proofs and computes the corresponding matching ELK expressions
such that all inferences can be lifted.