Fix missing inferences due to equivalences indexed as concept inclusions
(fix the problem exposed by the previous tests fixes)

If there was an equivalence C = D where both concepts are complex
or multiple equivalences with the same atomic concepts
(e.g., A = C and A = D), equivalences were converted to 
concept inclusions.
Hence inferences with these concepts inclusions did not have
equivalence axioms as premises, which resulted in missing proofs.

Now, in such cases, equivalence axioms are indexed as 
(binary) equivalence axioms which are used in dedicated inferences