Add inferences for special cases like singleton or empty conjunction
A singleton conjunction ObjectIntersectionOf(:D) is indexed by the
same expression as :D, so if SubClassInclusion matches, say
SubClassOf(:C ObjectIntersectionOf(:D)), it should be first
converted to SubClassOf(:C :D) using an appropriate inference.
This process of "canonizing" of conclusion matches is handled
by ConclusionMatchCanonizerVisitor.

Similar canonization is applied for subsumer matches that are 
singleton or empty ObjectUnionOf and ObjectOneOf.

Corresponding tests are added.