Fix problem with transitive reduction of partially supported class expressions
Some class expressions (disjunctions or OneOf) are partially supported, which
means that not all of their subsumers are derived. The optimization of
transitive reduction where equivalence is rested by comparing number of
subsumers would not work in this case. Now, this optimization is just a partial
filtering and equivalence is checked also by checking mutual subsumption.