Support deserialization
Deserializable returns a Deserializer. Each implementing class has inner
Deserializer that is able to call the constructor. Each Deserializer has a
unique ID and is registered in a common place. The first number on the input
stream needs to be the ID of the Deserializer that is able to construct its
object from the following data. If the number is not an ID of any registered
Deserializer, it is an error. If a different Deserializer was called and the
object cannot be cast to what it was supposed, it is an error.