Fix & simplify dynamic update of proofs. Test added.
It was wrong to assume that the proofs only change when
the tracing inference set is changed or the internal reasoner
of ElkReasoner changes. It may be, e.g., that the conclusion
was not derivable, but then becomes derivable, in which case
no changes were detected.

Now only one listener is used which monitors if the set of axioms
from which the entailments are computed have changed (e.g.,
after flush() is called)