Embed provided dependencies in elk-protege bundle
All elk dependencies that are embedded in Protege bundle are now
marked with the provided scope

This way they are not transitively inherited when elk-protege
is used as a dependencies

The embedded classes are now unpacked so that they can be accessed
not only from osgi (which uses a custom class loader) but also
from the ordinary applications. In other words there are no longer
jars within jars.