Fix concurrent computation test
It was false to assume that when each worker finishes the shared job
queue is empty. It may be that the last job is taken by one worker, the
other worker has nothing to process so it runs finish(), but the first
worker  puts the job to the shared queue, so it is not empty.