Ensure persistence of test names by changing sets to lists
Tests like ElkEntailmentQueryTest use a collection of input parameters.
The changes make sure that the order of elements in these parameters is
the same in each run. If the order is different, the test cannot be
re-run, e.g., from Eclipse, as the test with this name might not be
found at the next run.