Use features to detect trivial satisfiability of ontology; add tests
Some constructors, such as negation, did not result in indexing of
positive occurrences of owl:Nothing, which could result in incorrect
(trivial) satisfiability result. Now check explicitly for the presence
of constructors that can cause unsatisfiability.

Fix proof conversion for different individuals axiom that was revealed
by one of the tests.