Decouple puli from elk-reasoner
elk-reasoner no longer depends on puli and uses its own
version of inferences and proofs

This is done so that the reasoner implementation for the
owl-api interface does not depend on additional
(optional) libraries puli and owlapi-proofs, which are
provided by proof-based explanation plugins.

The problem was previously solved by providing these
additional libraries in the elk-protege zip, however
Protege currently installs only one (first) bundle from
a zip file.

Make puli and owlapi-proofs an optional dependency of
elk-protege for these reasons.